Organize the inheritance relationship to make the code at each level more streamlined.
