+++ Android/APIExample/app/src/main/java/io/agora/api/example/examples/basic_video_audio/AudioOnly.java
@@ -29,21 +29,9 @@ import io.agora.rtc.RtcEngine;
         name = "AudioOnly",
         actionId = R.id.action_mainFragment_to_audioOnly
 )
-public class AudioOnly extends BaseFragment implements View.OnClickListener
+public class AudioOnly extends JoinChannel implements View.OnClickListener
 {
-    private Button join, mute, speaker;
-    private EditText et_channel;
-
-    private Handler handler;
-    protected RtcEngine engine;
-    protected int myUid;
-
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState)
-    {
-        super.onCreate(savedInstanceState);
-        handler = new Handler();
-    }
+    private Button mute, speaker;
 
     @Nullable
     @Override
@@ -54,87 +42,21 @@ public class AudioOnly extends BaseFragment implements View.OnClickListener
         return view;
     }
 
+    @Override
     protected void initView(View view)
     {
-        join = view.findViewById(R.id.btn_join);
+        super.initView(view);
         mute = view.findViewById(R.id.btn_mute);
+        mute.setOnClickListener(this);
         speaker = view.findViewById(R.id.btn_speaker);
-        et_channel = view.findViewById(R.id.et_channel);
-        view.findViewById(R.id.btn_join).setOnClickListener(this);
-    }
-
-    @Override
-    public void onDestroy()
-    {
-        super.onDestroy();
-        // release the RtcEngine
-        engine.leaveChannel();
-        handler.post(RtcEngine::destroy);
-        engine = null;
-    }
-
-    protected void joinChannel(String channelId)
-    {
-        Context context = getContext();
-        if (context == null) return;
-
-        try
-        {
-            engine = RtcEngine.create(getContext(), getString(R.string.agora_app_id), eventHandler);
-            /** Sets the channel profile of the Agora RtcEngine.
-             CHANNEL_PROFILE_COMMUNICATION(0): (Default) The Communication profile.
-             Use this profile in one-on-one calls or group calls, where all users can talk freely.
-             CHANNEL_PROFILE_LIVE_BROADCASTING(1): The Live-Broadcast profile. Users in a live-broadcast
-             channel have a role as either broadcaster or audience. A broadcaster can both send and receive streams;
-             an audience can only receive streams.*/
-            engine.setChannelProfile(Constants.CHANNEL_PROFILE_COMMUNICATION);
-        }
-        catch (Exception e)
-        {
-            e.printStackTrace();
-        }
-
-        /** Allows a user to join a channel.
-         if you do not specify the uid, we will generate the uid for you*/
-        int res = engine.joinChannel(getString(R.string.agora_access_token), channelId, "Extra Optional Data", 0);
-        if (res != 0)
-        {
-            // Usually happens with invalid parameters
-            // Error code description can be found at:
-            // en: https://docs.agora.io/en/Voice/API%20Reference/java/classio_1_1agora_1_1rtc_1_1_i_rtc_engine_event_handler_1_1_error_code.html
-            // cn: https://docs.agora.io/cn/Voice/API%20Reference/java/classio_1_1agora_1_1rtc_1_1_i_rtc_engine_event_handler_1_1_error_code.html
-            showAlert(RtcEngine.getErrorDescription(Math.abs(res)));
-            return;
-        }
-        // Prevent repeated entry
-        join.setClickable(false);
+        speaker.setOnClickListener(this);
     }
 
     @Override
     public void onClick(View v)
     {
-        if (v.getId() == R.id.btn_join)
-        {
-            // call when join button hit
-            String channelId = et_channel.getText().toString();
-            // Check permission
-            if (AndPermission.hasPermissions(this, Permission.Group.STORAGE, Permission.Group.MICROPHONE, Permission.Group.CAMERA))
-            {
-                joinChannel(channelId);
-                return;
-            }
-            // Request permission
-            AndPermission.with(this).runtime().permission(
-                    Permission.Group.STORAGE,
-                    Permission.Group.MICROPHONE,
-                    Permission.Group.CAMERA
-            ).onGranted(permissions ->
-            {
-                // Permissions Granted
-                joinChannel(channelId);
-            }).start();
-        }
-        else if(v.getId() == R.id.btn_mute)
+        super.onClick(v);
+        if(v.getId() == R.id.btn_mute)
         {
             mute.setActivated(!mute.isActivated());
             mute.setText(getString(mute.isActivated() ? R.string.openmicrophone : R.string.closemicrophone));
@@ -148,115 +70,5 @@ public class AudioOnly extends BaseFragment implements View.OnClickListener
             // Turn off / on the speaker and change the audio playback route.
             engine.setEnableSpeakerphone(speaker.isActivated());
         }
-    }
-
-    private IRtcEngineEventHandler eventHandler = new IRtcEngineEventHandler()
-    {
-        private final String TAG = IRtcEngineEventHandler.class.getSimpleName();
-
-        @Override
-        public void onWarning(int warn)
-        {
-            Log.w(TAG, String.format("onWarning code %d message %s", warn, RtcEngine.getErrorDescription(warn)));
-        }
-
-        @Override
-        public void onError(int err)
-        {
-            Log.e(TAG, String.format("onError code %d message %s", err, RtcEngine.getErrorDescription(err)));
-        }
-
-        @Override
-        public void onJoinChannelSuccess(String channel, int uid, int elapsed)
-        {
-            Log.i(TAG, String.format("onJoinChannelSuccess channel %s uid %d", channel, uid));
-            showLongToast(String.format("onJoinChannelSuccess channel %s uid %d", channel, uid));
-            myUid = uid;
-            /** After successfully joining the channel, the user is allowed to switch
-             * the audio route and switch the microphone.*/
-            mute.setOnClickListener(AudioOnly.this);
-            speaker.setOnClickListener(AudioOnly.this);
-        }
-
-        /**Since v2.9.0.
-         * This callback indicates the state change of the remote audio stream.
-         * PS: This callback does not work properly when the number of users (in the Communication profile) or
-         *     broadcasters (in the Live-broadcast profile) in the channel exceeds 17.
-         * @param uid ID of the user whose audio state changes.
-         * @param state State of the remote audio
-         *   REMOTE_AUDIO_STATE_STOPPED(0): The remote audio is in the default state, probably due
-         *              to REMOTE_AUDIO_REASON_LOCAL_MUTED(3), REMOTE_AUDIO_REASON_REMOTE_MUTED(5),
-         *              or REMOTE_AUDIO_REASON_REMOTE_OFFLINE(7).
-         *   REMOTE_AUDIO_STATE_STARTING(1): The first remote audio packet is received.
-         *   REMOTE_AUDIO_STATE_DECODING(2): The remote audio stream is decoded and plays normally,
-         *              probably due to REMOTE_AUDIO_REASON_NETWORK_RECOVERY(2),
-         *              REMOTE_AUDIO_REASON_LOCAL_UNMUTED(4) or REMOTE_AUDIO_REASON_REMOTE_UNMUTED(6).
-         *   REMOTE_AUDIO_STATE_FROZEN(3): The remote audio is frozen, probably due to
-         *              REMOTE_AUDIO_REASON_NETWORK_CONGESTION(1).
-         *   REMOTE_AUDIO_STATE_FAILED(4): The remote audio fails to start, probably due to
-         *              REMOTE_AUDIO_REASON_INTERNAL(0).
-         * @param reason The reason of the remote audio state change.
-         *   REMOTE_AUDIO_REASON_INTERNAL(0): Internal reasons.
-         *   REMOTE_AUDIO_REASON_NETWORK_CONGESTION(1): Network congestion.
-         *   REMOTE_AUDIO_REASON_NETWORK_RECOVERY(2): Network recovery.
-         *   REMOTE_AUDIO_REASON_LOCAL_MUTED(3): The local user stops receiving the remote audio
-         *               stream or disables the audio module.
-         *   REMOTE_AUDIO_REASON_LOCAL_UNMUTED(4): The local user resumes receiving the remote audio
-         *              stream or enables the audio module.
-         *   REMOTE_AUDIO_REASON_REMOTE_MUTED(5): The remote user stops sending the audio stream or
-         *               disables the audio module.
-         *   REMOTE_AUDIO_REASON_REMOTE_UNMUTED(6): The remote user resumes sending the audio stream
-         *              or enables the audio module.
-         *   REMOTE_AUDIO_REASON_REMOTE_OFFLINE(7): The remote user leaves the channel.
-         *   @param elapsed Time elapsed (ms) from the local user calling the joinChannel method
-         *                  until the SDK triggers this callback.*/
-        @Override
-        public void onRemoteAudioStateChanged(int uid, int state, int reason, int elapsed)
-        {
-            super.onRemoteAudioStateChanged(uid, state, reason, elapsed);
-            Log.e(TAG, "onRemoteAudioStateChanged->" + uid + ", state->" + state + ", reason->" + reason);
-        }
-
-        /**Occurs when a remote user (Communication)/host (Live Broadcast) joins the channel.
-         * @param uid ID of the user whose audio state changes.
-         * @param elapsed Time delay (ms) from the local user calling joinChannel/setClientRole
-         *                until this callback is triggered.*/
-        @Override
-        public void onUserJoined(int uid, int elapsed)
-        {
-            super.onUserJoined(uid, elapsed);
-            Log.e(TAG, "onUserJoined->" + uid);
-            showLongToast(String.format("user %d joined!", uid));
-        }
-
-        /**Occurs when a remote user (Communication)/host (Live Broadcast) leaves the channel.
-         * @param uid ID of the user whose audio state changes.
-         * @param reason Reason why the user goes offline:
-         *   USER_OFFLINE_QUIT(0): The user left the current channel.
-         *   USER_OFFLINE_DROPPED(1): The SDK timed out and the user dropped offline because no data
-         *              packet was received within a certain period of time. If a user quits the
-         *               call and the message is not passed to the SDK (due to an unreliable channel),
-         *               the SDK assumes the user dropped offline.
-         *   USER_OFFLINE_BECOME_AUDIENCE(2): (Live broadcast only.) The client role switched from
-         *               the host to the audience.*/
-        @Override
-        public void onUserOffline(int uid, int reason)
-        {
-            Log.i(TAG, String.format("user %d offline! reason:%d", uid, reason));
-            showLongToast(String.format("user %d offline! reason:%d", uid, reason));
-        }
-    };
-    private final void showLongToast(final String msg)
-    {
-        if(getActivity() == null)
-        {return;}
-        getActivity().runOnUiThread(new Runnable()
-        {
-            @Override
-            public void run()
-            {
-                Toast.makeText(getContext().getApplicationContext(), msg, Toast.LENGTH_LONG).show();
-            }
-        });
     }
 }
