+++ Android/APIExample/app/src/main/java/io/agora/api/example/examples/basic_video_audio/JoinChannel.java
@@ -4,12 +4,13 @@ import android.content.Context;
 import android.os.Bundle;
 import android.os.Handler;
 import android.util.Log;
+import android.view.KeyEvent;
 import android.view.LayoutInflater;
-import android.view.SurfaceView;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
 import android.widget.EditText;
-import android.widget.FrameLayout;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -20,185 +21,273 @@ import com.yanzhenjie.permission.runtime.Permission;
 import io.agora.api.example.R;
 import io.agora.api.example.annotation.Example;
 import io.agora.api.example.common.BaseFragment;
+import io.agora.rtc.Constants;
+import io.agora.rtc.IRtcChannelEventHandler;
 import io.agora.rtc.IRtcEngineEventHandler;
 import io.agora.rtc.RtcEngine;
-import io.agora.rtc.video.VideoCanvas;
-import io.agora.rtc.video.VideoEncoderConfiguration;
-
-import static io.agora.rtc.video.VideoCanvas.RENDER_MODE_HIDDEN;
-import static io.agora.rtc.video.VideoEncoderConfiguration.FRAME_RATE.FRAME_RATE_FPS_15;
-import static io.agora.rtc.video.VideoEncoderConfiguration.ORIENTATION_MODE.ORIENTATION_MODE_ADAPTIVE;
-import static io.agora.rtc.video.VideoEncoderConfiguration.STANDARD_BITRATE;
-import static io.agora.rtc.video.VideoEncoderConfiguration.VD_640x360;
 
 @Example(
         group = "BASIC VIDEO/AUDIO",
         name = "JoinChannel",
         actionId = R.id.action_mainFragment_to_joinChannel
 )
-public class JoinChannel extends BaseFragment implements View.OnClickListener {
-    private FrameLayout fl_local, fl_remote;
+public class JoinChannel extends BaseFragment implements View.OnClickListener
+{
+    private static final String TAG = JoinChannel.class.getSimpleName();
+    private Button join;
     private EditText et_channel;
-
-    private Handler handler;
+    protected Handler handler;
     protected RtcEngine engine;
     protected int myUid;
+    private boolean joined = false;
 
     @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
+    public void onCreate(@Nullable Bundle savedInstanceState)
+    {
         super.onCreate(savedInstanceState);
         handler = new Handler();
     }
 
     @Nullable
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState)
+    {
         View view = inflater.inflate(R.layout.fragment_join_channel, container, false);
         initView(view);
         return view;
     }
 
-    protected void initView(View view) {
-        fl_local = view.findViewById(R.id.fl_local);
-        fl_remote = view.findViewById(R.id.fl_remote);
+    @Override
+    public void onActivityCreated(@Nullable Bundle savedInstanceState)
+    {
+        super.onActivityCreated(savedInstanceState);
+        initEngine();
+    }
+
+    protected void initView(View view)
+    {
+        join = view.findViewById(R.id.btn_join);
         et_channel = view.findViewById(R.id.et_channel);
         view.findViewById(R.id.btn_join).setOnClickListener(this);
     }
 
     @Override
-    public void onDestroy() {
+    public void onDestroy()
+    {
         super.onDestroy();
         // release the RtcEngine
+        if(engine != null)
+        {
+            engine.leaveChannel();
+        }
         handler.post(RtcEngine::destroy);
         engine = null;
     }
 
-    protected void joinChannel(String channelId) {
+    private void initEngine()
+    {
+        // Check if the context is valid
         Context context = getContext();
-        if (context == null) return;
+        if (context == null)
+        {return;}
 
-        try {
-            // 1. Create the RtcEngine with Agora app id
-            // Note: Please use ApplicationContext instead of ActivityContext
-            engine = RtcEngine.create(context.getApplicationContext(), getString(R.string.agora_app_id), eventHandler);
-        } catch (Exception e) {
-            e.printStackTrace();
+        try
+        {
+            engine = RtcEngine.create(getContext().getApplicationContext(), getString(R.string.agora_app_id), getEventHandler());
+            /** Sets the channel profile of the Agora RtcEngine.
+             CHANNEL_PROFILE_COMMUNICATION(0): (Default) The Communication profile.
+             Use this profile in one-on-one calls or group calls, where all users can talk freely.
+             CHANNEL_PROFILE_LIVE_BROADCASTING(1): The Live-Broadcast profile. Users in a live-broadcast
+             channel have a role as either broadcaster or audience. A broadcaster can both send and receive streams;
+             an audience can only receive streams.*/
+            engine.setChannelProfile(Constants.CHANNEL_PROFILE_COMMUNICATION);
         }
+        catch (Exception e)
+        {
+            e.printStackTrace();
+            getActivity().onBackPressed();
+        }
+    }
 
-        // Enable video module
-        engine.enableVideo();
-        // Setup video encoding configs
-        engine.setVideoEncoderConfiguration(new VideoEncoderConfiguration(
-                VD_640x360,
-                FRAME_RATE_FPS_15,
-                STANDARD_BITRATE,
-                ORIENTATION_MODE_ADAPTIVE
-        ));
-
-        // Create render view by RtcEngine
-        SurfaceView surfaceView = RtcEngine.CreateRendererView(context);
-        // Local video is on the top
-        surfaceView.setZOrderMediaOverlay(true);
-        // Add to the local container
-        fl_local.addView(surfaceView, new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
-        // Setup local video to render your local camera preview
-        engine.setupLocalVideo(new VideoCanvas(surfaceView, RENDER_MODE_HIDDEN, 0));
-
-        // Set audio route to speaker
-        engine.setDefaultAudioRoutetoSpeakerphone(true);
-
-        // Start joining channel
-        // 1. Users can only see each other after they join the same channel successfully using the same app id
-        // 2. If app certificate is turned on at dashboard, token is needed when joining channel
-        // The channel name and uid used to calculate the token has to match the ones used for channel join
-        int res = engine.joinChannel(getString(R.string.agora_access_token), channelId, null, 0);
-        if (res != 0) {
+    protected void joinChannel(String channelId)
+    {
+        /** Allows a user to join a channel.
+         if you do not specify the uid, we will generate the uid for you*/
+        int res = engine.joinChannel(getString(R.string.agora_access_token), channelId, "Extra Optional Data", 0);
+        if (res != 0)
+        {
             // Usually happens with invalid parameters
             // Error code description can be found at:
             // en: https://docs.agora.io/en/Voice/API%20Reference/java/classio_1_1agora_1_1rtc_1_1_i_rtc_engine_event_handler_1_1_error_code.html
             // cn: https://docs.agora.io/cn/Voice/API%20Reference/java/classio_1_1agora_1_1rtc_1_1_i_rtc_engine_event_handler_1_1_error_code.html
             showAlert(RtcEngine.getErrorDescription(Math.abs(res)));
+            return;
         }
+        // Prevent repeated entry
+        join.setClickable(false);
     }
 
     @Override
-    public void onClick(View v) {
-        if (v.getId() == R.id.btn_join) { // call when join button hit
-            String channelId = et_channel.getText().toString();
-
-            // Check permission
-            if (AndPermission.hasPermissions(this, Permission.Group.STORAGE, Permission.Group.MICROPHONE, Permission.Group.CAMERA)) {
-                joinChannel(channelId);
-                return;
+    public void onClick(View v)
+    {
+        if (v.getId() == R.id.btn_join)
+        {
+            if(!joined)
+            {
+                // call when join button hit
+                String channelId = et_channel.getText().toString();
+                // Check permission
+                if (AndPermission.hasPermissions(this, Permission.Group.STORAGE, Permission.Group.MICROPHONE, Permission.Group.CAMERA))
+                {
+                    joinChannel(channelId);
+                    return;
+                }
+                // Request permission
+                AndPermission.with(this).runtime().permission(
+                        Permission.Group.STORAGE,
+                        Permission.Group.MICROPHONE,
+                        Permission.Group.CAMERA
+                ).onGranted(permissions ->
+                {
+                    // Permissions Granted
+                    joinChannel(channelId);
+                }).start();
             }
-            // Request permission
-            AndPermission.with(this).runtime().permission(
-                    Permission.Group.STORAGE,
-                    Permission.Group.MICROPHONE,
-                    Permission.Group.CAMERA
-            ).onGranted(permissions -> { // Permissions Granted
-                joinChannel(channelId);
-            }).start();
+            else
+            {
+                joined = false;
+                engine.leaveChannel();
+                join.setText(getString(R.string.join));
+            }
         }
     }
 
-    private IRtcEngineEventHandler eventHandler = new IRtcEngineEventHandler() {
+    protected IRtcEngineEventHandler getEventHandler()
+    {return joinChannelEventHandler;}
+
+    protected IRtcEngineEventHandler joinChannelEventHandler = new IRtcEngineEventHandler()
+    {
         private final String TAG = IRtcEngineEventHandler.class.getSimpleName();
 
         @Override
-        public void onWarning(int warn) {
+        public void onWarning(int warn)
+        {
             Log.w(TAG, String.format("onWarning code %d message %s", warn, RtcEngine.getErrorDescription(warn)));
         }
 
         @Override
-        public void onError(int err) {
+        public void onError(int err)
+        {
             Log.e(TAG, String.format("onError code %d message %s", err, RtcEngine.getErrorDescription(err)));
         }
 
+        /**Occurs when a user leaves the channel.
+         * @param stats With this callback, the application retrieves the channel information,
+         *              such as the call duration and statistics.*/
         @Override
-        public void onJoinChannelSuccess(String channel, int uid, int elapsed) {
+        public void onLeaveChannel(RtcStats stats)
+        {
+            super.onLeaveChannel(stats);
+            showLongToast(String.format("local user %d leaveChannel!", myUid));
+        }
+
+        @Override
+        public void onJoinChannelSuccess(String channel, int uid, int elapsed)
+        {
             Log.i(TAG, String.format("onJoinChannelSuccess channel %s uid %d", channel, uid));
+            showLongToast(String.format("onJoinChannelSuccess channel %s uid %d", channel, uid));
             myUid = uid;
-        }
-
-        @Override
-        public void onLeaveChannel(RtcStats stats) {
-            Log.i(TAG, "onLeaveChannel");
-        }
-
-        @Override
-        public void onUserJoined(int uid, int elapsed) {
-            Log.i(TAG, String.format("onUserJoined uid %d", uid));
-
-            Context context = getContext();
-            if (context == null) return;
-
-            handler.post(() -> {
-                SurfaceView surfaceView = null;
-                if (fl_remote.getChildCount() == 0) {
-                    // Create render view by RtcEngine
-                    surfaceView = RtcEngine.CreateRendererView(context);
-                    // Add to the remote container
-                    fl_remote.addView(surfaceView, new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
-                } else {
-                    View view = fl_remote.getChildAt(0);
-                    if (view instanceof SurfaceView) {
-                        surfaceView = (SurfaceView) view;
-                    }
+            joined = true;
+            handler.post(new Runnable()
+            {
+                @Override
+                public void run()
+                {
+                    join.setClickable(true);
+                    join.setText(getString(R.string.leave));
                 }
-
-                // Setup remote video to render
-                engine.setupRemoteVideo(new VideoCanvas(surfaceView, RENDER_MODE_HIDDEN, uid));
             });
         }
 
+        /**Since v2.9.0.
+         * This callback indicates the state change of the remote audio stream.
+         * PS: This callback does not work properly when the number of users (in the Communication profile) or
+         *     broadcasters (in the Live-broadcast profile) in the channel exceeds 17.
+         * @param uid ID of the user whose audio state changes.
+         * @param state State of the remote audio
+         *   REMOTE_AUDIO_STATE_STOPPED(0): The remote audio is in the default state, probably due
+         *              to REMOTE_AUDIO_REASON_LOCAL_MUTED(3), REMOTE_AUDIO_REASON_REMOTE_MUTED(5),
+         *              or REMOTE_AUDIO_REASON_REMOTE_OFFLINE(7).
+         *   REMOTE_AUDIO_STATE_STARTING(1): The first remote audio packet is received.
+         *   REMOTE_AUDIO_STATE_DECODING(2): The remote audio stream is decoded and plays normally,
+         *              probably due to REMOTE_AUDIO_REASON_NETWORK_RECOVERY(2),
+         *              REMOTE_AUDIO_REASON_LOCAL_UNMUTED(4) or REMOTE_AUDIO_REASON_REMOTE_UNMUTED(6).
+         *   REMOTE_AUDIO_STATE_FROZEN(3): The remote audio is frozen, probably due to
+         *              REMOTE_AUDIO_REASON_NETWORK_CONGESTION(1).
+         *   REMOTE_AUDIO_STATE_FAILED(4): The remote audio fails to start, probably due to
+         *              REMOTE_AUDIO_REASON_INTERNAL(0).
+         * @param reason The reason of the remote audio state change.
+         *   REMOTE_AUDIO_REASON_INTERNAL(0): Internal reasons.
+         *   REMOTE_AUDIO_REASON_NETWORK_CONGESTION(1): Network congestion.
+         *   REMOTE_AUDIO_REASON_NETWORK_RECOVERY(2): Network recovery.
+         *   REMOTE_AUDIO_REASON_LOCAL_MUTED(3): The local user stops receiving the remote audio
+         *               stream or disables the audio module.
+         *   REMOTE_AUDIO_REASON_LOCAL_UNMUTED(4): The local user resumes receiving the remote audio
+         *              stream or enables the audio module.
+         *   REMOTE_AUDIO_REASON_REMOTE_MUTED(5): The remote user stops sending the audio stream or
+         *               disables the audio module.
+         *   REMOTE_AUDIO_REASON_REMOTE_UNMUTED(6): The remote user resumes sending the audio stream
+         *              or enables the audio module.
+         *   REMOTE_AUDIO_REASON_REMOTE_OFFLINE(7): The remote user leaves the channel.
+         *   @param elapsed Time elapsed (ms) from the local user calling the joinChannel method
+         *                  until the SDK triggers this callback.*/
         @Override
-        public void onUserOffline(int uid, int reason) {
-            Log.i(TAG, String.format("onUserOffline uid %d reason %d", uid, reason));
+        public void onRemoteAudioStateChanged(int uid, int state, int reason, int elapsed)
+        {
+            super.onRemoteAudioStateChanged(uid, state, reason, elapsed);
+            Log.e(TAG, "onRemoteAudioStateChanged->" + uid + ", state->" + state + ", reason->" + reason);
+        }
 
-            // Clear render view
-            // Note: The video will stay at its last frame, to completely remove it you will need to remove the SurfaceView from its parent
-            engine.setupRemoteVideo(new VideoCanvas(null, RENDER_MODE_HIDDEN, uid));
+        /**Occurs when a remote user (Communication)/host (Live Broadcast) joins the channel.
+         * @param uid ID of the user whose audio state changes.
+         * @param elapsed Time delay (ms) from the local user calling joinChannel/setClientRole
+         *                until this callback is triggered.*/
+        @Override
+        public void onUserJoined(int uid, int elapsed)
+        {
+            super.onUserJoined(uid, elapsed);
+            Log.e(TAG, "onUserJoined->" + uid);
+            showLongToast(String.format("user %d joined!", uid));
+        }
+
+        /**Occurs when a remote user (Communication)/host (Live Broadcast) leaves the channel.
+         * @param uid ID of the user whose audio state changes.
+         * @param reason Reason why the user goes offline:
+         *   USER_OFFLINE_QUIT(0): The user left the current channel.
+         *   USER_OFFLINE_DROPPED(1): The SDK timed out and the user dropped offline because no data
+         *              packet was received within a certain period of time. If a user quits the
+         *               call and the message is not passed to the SDK (due to an unreliable channel),
+         *               the SDK assumes the user dropped offline.
+         *   USER_OFFLINE_BECOME_AUDIENCE(2): (Live broadcast only.) The client role switched from
+         *               the host to the audience.*/
+        @Override
+        public void onUserOffline(int uid, int reason)
+        {
+            Log.i(TAG, String.format("user %d offline! reason:%d", uid, reason));
+            showLongToast(String.format("user %d offline! reason:%d", uid, reason));
         }
     };
+    private final void showLongToast(final String msg)
+    {
+        handler.post(new Runnable()
+        {
+            @Override
+            public void run()
+            {
+                if (JoinChannel.this == null || getContext() == null)
+                {return;}
+                Toast.makeText(getContext().getApplicationContext(), msg, Toast.LENGTH_LONG).show();
+            }
+        });
+    }
 }
