+++ Android/APIExample/app/src/main/res/layout/fragment_join_channel.xml
@@ -3,20 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".examples.basic_video_audio.JoinChannel">
-
-    <FrameLayout
-        android:id="@+id/fl_remote"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_above="@+id/ll_join" />
-
-    <FrameLayout
-        android:id="@+id/fl_local"
-        android:layout_width="150dp"
-        android:layout_height="200dp"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true" />
+    tools:context=".examples.basic_video_audio.AudioOnly">
 
     <LinearLayout
         android:id="@+id/ll_join"
